requirements.txt
streamlit
pyttsx3
requests
pyautogui
pywhatkit
google-generativeai
